// src/features/News/types/queryTypes.ts
export interface ApiResponse<T> {
  data: T;
  pagination?: {
    total: number;
    pages: number;
  };
}


//Paginations
interface PaginationByPage {
  page?: number;
  pageSize?: number;
}

interface PaginationByOffset {
  start?: number;
  limit?: number;
}

type StrapiPagination = 
  | { pagination: PaginationByPage }
  | { pagination: PaginationByOffset };

//   Sort
  type SortOrder = 'asc' | 'desc';

interface SortParams {
  sort?: string | string[] | { [field: string]: SortOrder };
}

// Filter
type FilterOperator =
  | '$eq'
  | '$ne'
  | '$lt'
  | '$lte'
  | '$gt'
  | '$gte'
  | '$in'
  | '$notIn'
  | '$contains'
  | '$notContains'
  | '$containsi'
  | '$notContainsi'
  | '$null'
  | '$notNull'
  | '$between'
  | '$startsWith'
  | '$endsWith';

type FilterValue = string | number | boolean | Date | Array<string | number> | null;

interface StrapiFilter {
  [field: string]:
    | string 
    | number 
    | { [operator in FilterOperator]?: FilterValue }
    | { [operator in FilterOperator]?: [FilterValue, FilterValue] } // For between
    | StrapiFilter // For nested relations
    | Array<StrapiFilter>; // For $and/$or
}

interface filterBasis {
  filters?: StrapiFilter;
}

//---- 
interface FieldsParams {
  fields?: string[] | { [model: string]: string[] };
}

//----
interface PopulateParams {
  populate?:
    | string
    | string[]
    | Record<string,string | boolean>
    | { [relation: string]: { fields?: string[]; filters?: StrapiFilter } };
}
//--- 
export type backendQueryParams = 
  Partial<StrapiPagination> &
  Partial<filterBasis> &
  Partial<SortParams> &
  Partial<FieldsParams> &
  Partial<PopulateParams> & {
    publicationState?: 'live' | 'preview';
    locale?: string | string[];
  };